// src/constants/enums.ts
enum ScreenEnum {
  DIRECTORY = 'Directory',
  EDITOR = 'Editor',
  MARKUP = 'Markup',
  PRIVACY_POLICY = 'Privacy Policy',
  TERMS_OF_SERVICE = 'Terms of Service',
}

enum EntityEnum {
  CLIPBOARD = 'clipboard',
  PROJECT = 'project',
  MARKUP = 'markup',
  WEBPAGE = 'webpage',
  NOTE = 'note',
  URL = 'url', // just the link, and AI summarizes
}

export { ScreenEnum, EntityEnum };